[CAMEL-11767] Create camel-tooling-maven module that provides Maven downloader for camel-catalog-maven and camel-kamelet-main
