+++ bom/camel-bom/pom.xml
@@ -1874,6 +1874,11 @@
       </dependency>
       <dependency>
         <groupId>org.apache.camel</groupId>
+        <artifactId>camel-tooling-maven</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.camel</groupId>
         <artifactId>camel-tooling-model</artifactId>
         <version>${project.version}</version>
       </dependency>
