+++ catalog/camel-catalog-maven/src/main/java/org/apache/camel/catalog/maven/MavenVersionManager.java
@@ -19,15 +19,22 @@ package org.apache.camel.catalog.maven;
 import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.Collections;
 import java.util.Enumeration;
-import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.LinkedHashSet;
+import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
-import groovy.grape.Grape;
-import groovy.lang.GroovyClassLoader;
 import org.apache.camel.catalog.VersionManager;
-import org.apache.ivy.util.url.URLHandlerRegistry;
+import org.apache.camel.tooling.maven.MavenArtifact;
+import org.apache.camel.tooling.maven.MavenDownloader;
+import org.apache.camel.tooling.maven.MavenDownloaderImpl;
+import org.apache.camel.tooling.maven.MavenResolutionException;
+import org.eclipse.aether.ConfigurationProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,19 +42,33 @@ import org.slf4j.LoggerFactory;
  * A {@link VersionManager} that can load the resources using Maven to download needed artifacts from a local or remote
  * Maven repository.
  * <p/>
- * This implementation uses Groovy Grape to download the Maven JARs.
+ * This implementation uses Maven Resolver to download the Maven JARs.
  */
 public class MavenVersionManager implements VersionManager, Closeable {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MavenVersionManager.class);
 
     private ClassLoader classLoader;
-    private final ClassLoader groovyClassLoader = new GroovyClassLoader();
-    private final TimeoutHttpClientHandler httpClient = new TimeoutHttpClientHandler();
+    private final OpenURLClassLoader helperClassLoader = new OpenURLClassLoader();
+
     private String version;
     private String runtimeProviderVersion;
-    private String cacheDirectory;
+    private String localRepository;
     private boolean log;
+
+    private final MavenDownloader downloader;
+
+    private final Map<String, String> repositories = new LinkedHashMap<>();
+
+    private int connectTimeout = ConfigurationProperties.DEFAULT_CONNECT_TIMEOUT;
+    private int requestTimeout = ConfigurationProperties.DEFAULT_REQUEST_TIMEOUT;
+
+    private boolean customized;
+
+    public MavenVersionManager() {
+        downloader = new MavenDownloaderImpl();
+        ((MavenDownloaderImpl) downloader).build();
+    }
 
     @Override
     public void setClassLoader(ClassLoader classLoader) {
@@ -57,12 +78,13 @@ public class MavenVersionManager implements VersionManager, Closeable {
     /**
      * Configures the directory for the download cache.
      * <p/>
-     * The default folder is <tt>USER_HOME/.groovy/grape</tt>
+     * The default folder is <tt>USER_HOME/.m2/repository</tt>
      *
      * @param directory the directory.
      */
     public void setCacheDirectory(String directory) {
-        this.cacheDirectory = directory;
+        this.localRepository = directory;
+        this.customized = true;
     }
 
     /**
@@ -73,12 +95,23 @@ public class MavenVersionManager implements VersionManager, Closeable {
     }
 
     /**
-     * Sets the timeout in millis (http.socket.timeout) when downloading via http/https protocols.
+     * Sets the connection timeout in millis when downloading via http/https protocols.
      * <p/>
      * The default value is 10000
      */
     public void setHttpClientTimeout(int timeout) {
-        httpClient.setTimeout(timeout);
+        this.connectTimeout = timeout;
+        this.customized = true;
+    }
+
+    /**
+     * Sets the read timeout in millis when downloading via http/https protocols.
+     * <p/>
+     * The default value is 1800000
+     */
+    public void setHttpClientRequestTimeout(int timeout) {
+        this.requestTimeout = timeout;
+        this.customized = true;
     }
 
     /**
@@ -88,10 +121,7 @@ public class MavenVersionManager implements VersionManager, Closeable {
      * @param url  the repository url
      */
     public void addMavenRepository(String name, String url) {
-        Map<String, Object> repo = new HashMap<>();
-        repo.put("name", name);
-        repo.put("root", url);
-        Grape.addResolver(repo);
+        repositories.put(name, url);
     }
 
     @Override
@@ -102,21 +132,13 @@ public class MavenVersionManager implements VersionManager, Closeable {
     @Override
     public boolean loadVersion(String version) {
         try {
-            URLHandlerRegistry.setDefault(httpClient);
-
-            if (cacheDirectory != null) {
-                System.setProperty("grape.root", cacheDirectory);
+            MavenDownloader mavenDownloader = downloader;
+            if (customized) {
+                mavenDownloader = mavenDownloader.customize(localRepository, connectTimeout, requestTimeout);
             }
 
-            Grape.setEnableAutoDownload(true);
-
-            Map<String, Object> param = new HashMap<>();
-            param.put("classLoader", groovyClassLoader);
-            param.put("group", "org.apache.camel");
-            param.put("module", "camel-catalog");
-            param.put("version", version);
-
-            Grape.grab(param);
+            String camelCatalogGAV = String.format("org.apache.camel:camel-catalog:%s", version);
+            resolve(mavenDownloader, camelCatalogGAV, version.contains("SNAPSHOT"));
 
             this.version = version;
             return true;
@@ -136,17 +158,13 @@ public class MavenVersionManager implements VersionManager, Closeable {
     @Override
     public boolean loadRuntimeProviderVersion(String groupId, String artifactId, String version) {
         try {
-            URLHandlerRegistry.setDefault(httpClient);
+            MavenDownloader mavenDownloader = downloader;
+            if (customized) {
+                mavenDownloader = mavenDownloader.customize(localRepository, connectTimeout, requestTimeout);
+            }
 
-            Grape.setEnableAutoDownload(true);
-
-            Map<String, Object> param = new HashMap<>();
-            param.put("classLoader", groovyClassLoader);
-            param.put("group", groupId);
-            param.put("module", artifactId);
-            param.put("version", version);
-
-            Grape.grab(param);
+            String gav = String.format("%s:%s:%s", groupId, artifactId, version);
+            resolve(mavenDownloader, gav, version.contains("SNAPSHOT"));
 
             this.runtimeProviderVersion = version;
             return true;
@@ -155,6 +173,27 @@ public class MavenVersionManager implements VersionManager, Closeable {
                 LOGGER.warn("Cannot load runtime provider version {} due {}", version, e.getMessage(), e);
             }
             return false;
+        }
+    }
+
+    /**
+     * Resolves Maven artifact using passed coordinates and use downloaded artifact as one of the URLs in the
+     * helperClassLoader, so further Catalog access may load resources from it.
+     *
+     * @param mavenDownloader
+     * @param gav
+     * @param useSnapshots
+     */
+    private void resolve(MavenDownloader mavenDownloader, String gav, boolean useSnapshots)
+            throws MavenResolutionException, MalformedURLException {
+        Set<String> extraRepositories = new LinkedHashSet<>(repositories.values());
+
+        // non-transitive resolve, because we load static data from the catalog artifacts
+        List<MavenArtifact> artifacts = mavenDownloader.resolveArtifacts(Collections.singletonList(gav),
+                extraRepositories, false, useSnapshots);
+
+        for (MavenArtifact ma : artifacts) {
+            helperClassLoader.addURL(ma.getFile().toURI().toURL());
         }
     }
 
@@ -175,7 +214,7 @@ public class MavenVersionManager implements VersionManager, Closeable {
             is = MavenVersionManager.class.getClassLoader().getResourceAsStream(name);
         }
         if (is == null) {
-            is = groovyClassLoader.getResourceAsStream(name);
+            is = helperClassLoader.getResourceAsStream(name);
         }
 
         return is;
@@ -188,7 +227,7 @@ public class MavenVersionManager implements VersionManager, Closeable {
 
         try {
             URL found = null;
-            Enumeration<URL> urls = groovyClassLoader.getResources(name);
+            Enumeration<URL> urls = helperClassLoader.getResources(name);
             while (urls.hasMoreElements()) {
                 URL url = urls.nextElement();
                 if (url.getPath().contains(version)) {
