+++ dsl/camel-kamelet-main/src/test/java/org/apache/camel/main/MavenDependencyResolverTest.java
@@ -16,44 +16,15 @@
  */
 package org.apache.camel.main;
 
-import java.io.File;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.util.Base64;
 import java.util.List;
-import java.util.Properties;
-import java.util.concurrent.TimeUnit;
 
-import org.apache.camel.main.download.MavenArtifact;
 import org.apache.camel.main.download.MavenDependencyDownloader;
-import org.apache.camel.main.injection.DIRegistry;
-import org.apache.camel.util.FileUtil;
-import org.apache.commons.codec.binary.Hex;
-import org.apache.http.Header;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.bootstrap.HttpServer;
-import org.apache.http.impl.bootstrap.ServerBootstrap;
-import org.apache.maven.settings.Mirror;
-import org.apache.maven.settings.Repository;
-import org.apache.maven.settings.Settings;
-import org.eclipse.aether.RepositorySystem;
-import org.eclipse.aether.RepositorySystemSession;
-import org.eclipse.aether.artifact.DefaultArtifact;
-import org.eclipse.aether.repository.RemoteRepository;
-import org.eclipse.aether.resolution.ArtifactRequest;
-import org.eclipse.aether.resolution.ArtifactResult;
-import org.junit.jupiter.api.AfterAll;
+import org.apache.camel.tooling.maven.MavenArtifact;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.sonatype.plexus.components.cipher.DefaultPlexusCipher;
-import org.sonatype.plexus.components.cipher.PlexusCipher;
-import org.sonatype.plexus.components.sec.dispatcher.DefaultSecDispatcher;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -63,233 +34,27 @@ public class MavenDependencyResolverTest {
 
     public static final Logger LOG = LoggerFactory.getLogger(MavenDependencyResolverTest.class);
 
-    private static HttpServer localServer;
-
-    @BeforeAll
-    public static void startMavenMirror() throws Exception {
-        localServer = ServerBootstrap.bootstrap()
-                .setListenerPort(8888)
-                .registerHandler("/maven/*", (req, res, context) -> {
-                    Header authz = req.getFirstHeader("Authorization");
-                    if (authz == null) {
-                        res.addHeader("WWW-Authenticate", "Basic realm=Camel");
-                        res.setStatusCode(401);
-                        return;
-                    }
-                    String creds = new String(Base64.getDecoder().decode(authz.getValue().split(" ")[1]));
-                    if (!"camel:passw0rd".equals(creds)) {
-                        res.setStatusCode(403);
-                        return;
-                    }
-                    LOG.info("Request: {}", req.getRequestLine());
-                    String request = req.getRequestLine().getUri().substring("/maven".length());
-                    if (request.endsWith(".jar") || request.endsWith(".pom")) {
-                        res.setEntity(new StringEntity(request));
-                    } else {
-                        MessageDigest md = null;
-                        try {
-                            if (request.endsWith(".md5")) {
-                                md = MessageDigest.getInstance("MD5");
-                                request = request.substring(0, request.length() - 4);
-                            } else if (request.endsWith(".sha1")) {
-                                md = MessageDigest.getInstance("SHA");
-                                request = request.substring(0, request.length() - 5);
-                            }
-                            if (md != null) {
-                                byte[] digest = md.digest(request.getBytes(StandardCharsets.UTF_8));
-                                res.setEntity(new StringEntity(Hex.encodeHexString(digest)));
-                            }
-                        } catch (NoSuchAlgorithmException e) {
-                            throw new RuntimeException(e);
-                        }
-                    }
-                    res.setStatusCode(200);
-                })
-                .create();
-        localServer.start();
-    }
-
-    @AfterAll
-    public static void stopMavenMirror() {
-        if (localServer != null) {
-            localServer.shutdown(2, TimeUnit.SECONDS);
-        }
-    }
-
-    @Test
-    public void masterPasswords() throws Exception {
-        PlexusCipher pc = new DefaultPlexusCipher();
-        String mp = pc.encrypt("camel", DefaultSecDispatcher.SYSTEM_PROPERTY_SEC_LOCATION);
-        LOG.info("master password: {}", mp); // e.g., "SHrYKy0oCBEH5SHhRcuv0U52J3E908O23QWHDyEiGtQ="
-        String p = pc.encrypt("passw0rd", "camel");
-        LOG.info("password: {}", p); // e.g., "9V4tKIxO4ZsHx63bkn9uy6zsYM9VJyG03sTsPzPDK9c="
-
-        assertEquals("passw0rd", pc.decrypt(p, pc.decrypt(mp, DefaultSecDispatcher.SYSTEM_PROPERTY_SEC_LOCATION)));
-        assertEquals("camel", pc.decrypt(mp, DefaultSecDispatcher.SYSTEM_PROPERTY_SEC_LOCATION));
-    }
-
     @Test
     public void testDownload() throws Exception {
         List<String> deps = List.of("org.apache.camel:camel-core:3.17.0");
         try (MavenDependencyDownloader downloader = new MavenDependencyDownloader()) {
-            // build() creates own DIRegistry and configures the resolver/aether stuff
+            // build() creates own DIRegistry and configures the resolver/aether stuff  through
+            // org.apache.camel:camel-tooling-maven
             downloader.build();
 
-            List<MavenArtifact> answer = downloader.resolveDependenciesViaAether(deps, true);
+            List<MavenArtifact> answer = downloader.resolveDependenciesViaAether(deps, null, true, false);
             Assertions.assertNotNull(answer);
             assertTrue(answer.size() > 15);
             for (MavenArtifact ma : answer) {
                 LOG.info("Artifact (transitive): {}", ma);
             }
 
-            answer = downloader.resolveDependenciesViaAether(deps, false);
+            answer = downloader.resolveDependenciesViaAether(deps, null, false, false);
             Assertions.assertNotNull(answer);
             assertEquals(1, answer.size());
             for (MavenArtifact ma : answer) {
                 LOG.info("Artifact (non-transitive): {}", ma);
             }
-        }
-    }
-
-    @Test
-    public void playingWithRepositorySystem() throws Exception {
-        // MRESOLVER-157 is about removal of org.apache.maven.repository.internal.MavenRepositorySystemUtils,
-        // so we can 1) do it manually, 2) use guice/sisu
-
-        // org.eclipse.aether.RepositorySystem (2 levels down)
-        //    org.eclipse.aether.impl.VersionResolver
-        //       org.eclipse.aether.impl.MetadataResolver
-        //       org.eclipse.aether.impl.SyncContextFactory
-        //       org.eclipse.aether.impl.RepositoryEventDispatcher
-        //    org.eclipse.aether.impl.VersionRangeResolver
-        //       org.eclipse.aether.impl.MetadataResolver
-        //       org.eclipse.aether.impl.SyncContextFactory
-        //       org.eclipse.aether.impl.RepositoryEventDispatcher
-        //    org.eclipse.aether.impl.ArtifactResolver
-        //       org.eclipse.aether.spi.io.FileProcessor
-        //       org.eclipse.aether.impl.RepositoryEventDispatcher
-        //       org.eclipse.aether.impl.VersionResolver
-        //       org.eclipse.aether.impl.UpdateCheckManager
-        //       org.eclipse.aether.impl.RepositoryConnectorProvider
-        //       org.eclipse.aether.impl.RemoteRepositoryManager
-        //       org.eclipse.aether.spi.synccontext.SyncContextFactory
-        //       org.eclipse.aether.impl.OfflineController
-        //    org.eclipse.aether.impl.MetadataResolver
-        //       org.eclipse.aether.impl.RepositoryEventDispatcher
-        //       org.eclipse.aether.impl.UpdateCheckManager
-        //       org.eclipse.aether.impl.RepositoryConnectorProvider
-        //       org.eclipse.aether.impl.RemoteRepositoryManager
-        //       org.eclipse.aether.spi.synccontext.SyncContextFactory
-        //       org.eclipse.aether.impl.OfflineController
-        //    org.eclipse.aether.impl.ArtifactDescriptorReader
-        //       org.eclipse.aether.impl.RemoteRepositoryManager
-        //       org.eclipse.aether.impl.VersionResolver
-        //       org.eclipse.aether.impl.VersionRangeResolver
-        //       org.eclipse.aether.impl.ArtifactResolver
-        //       org.apache.maven.model.building.ModelBuilder
-        //       org.eclipse.aether.impl.RepositoryEventDispatcher
-        //    org.eclipse.aether.impl.DependencyCollector
-        //       Map<String, org.eclipse.aether.internal.impl.collect.DependencyCollectorDelegate>
-        //    org.eclipse.aether.impl.Installer
-        //       org.eclipse.aether.spi.io.FileProcessor
-        //       org.eclipse.aether.impl.RepositoryEventDispatcher
-        //       Set<org.eclipse.aether.impl.MetadataGeneratorFactory>
-        //       org.eclipse.aether.spi.synccontext.SyncContextFactory
-        //    org.eclipse.aether.impl.Deployer
-        //       org.eclipse.aether.spi.io.FileProcessor
-        //       org.eclipse.aether.impl.RepositoryEventDispatcher
-        //       org.eclipse.aether.impl.RepositoryConnectorProvider
-        //       org.eclipse.aether.impl.RemoteRepositoryManager
-        //       org.eclipse.aether.impl.UpdateCheckManager
-        //       Set<org.eclipse.aether.impl.MetadataGeneratorFactory>
-        //       org.eclipse.aether.spi.synccontext.SyncContextFactory
-        //       org.eclipse.aether.impl.OfflineController
-        //    org.eclipse.aether.impl.LocalRepositoryProvider
-        //       Set<org.eclipse.aether.spi.localrepo.LocalRepositoryManagerFactory>
-        //    org.eclipse.aether.spi.synccontext.SyncContextFactory
-        //       org.eclipse.aether.internal.impl.synccontext.named.NamedLockFactorySelector
-        //    org.eclipse.aether.impl.RemoteRepositoryManager
-        //       org.eclipse.aether.impl.UpdatePolicyAnalyzer
-        //       org.eclipse.aether.spi.connector.checksum.ChecksumPolicyProvider
-
-        try (DIRegistry registry = new DIRegistry();
-             MavenDependencyDownloader downloader = new MavenDependencyDownloader()) {
-            // here we don't call downloader.build() and will do the same stuff manually for demonstration purpose
-
-            // see org.eclipse.aether.impl.DefaultServiceLocator.DefaultServiceLocator() - it registers
-            // lots of default implementations to get started (but it's deprecated with MRESOLVER-157)
-
-            // global settings file is by default ${m2.home}/.m2/settings.xml
-            //  - no defaults in Maven, but shrinkwrap uses this order:
-            //     - -Dmaven.home
-            //     - $M2_HOME
-            //     - $MAVEN_HOME
-            // local settings file is by default ${user.home}/.m2/settings.xml
-            // security settings file is by default ~/.m2/settings-security.xml
-            // (see: org.sonatype.plexus.components.sec.dispatcher.DefaultSecDispatcher), but it may be altered
-            // but it may be altered by (also from plexus-sec-dispatcher) -Dsettings.security
-            // local repository is by default ${user.home}/.m2/repository, but may be altered by
-            // -Dmaven.repo.local
-            String localSettings = "target/test-classes/.m2/settings.xml";
-            String localSettingsSecurity = "target/test-classes/.m2/settings-security.xml";
-
-            downloader.setMavenSettings(localSettings);
-            downloader.setMavenSettingsSecurity(localSettingsSecurity);
-            downloader.setFresh(true);
-            downloader.setRepos(null);
-            //            downloader.build();
-
-            Properties systemProperties = new Properties();
-
-            // now, org.eclipse.aether.RepositorySystem can be obtained
-            RepositorySystem repositorySystem
-                    = downloader.configureRepositorySystem(registry, systemProperties, localSettingsSecurity);
-
-            Settings settings = downloader.mavenConfiguration(registry, repositorySystem,
-                    systemProperties, localSettings);
-
-            // when using Maven without a project that may contain <repositories> in pom.xml, repositories
-            // are taken from the active profiles defined in settings. If a repository is protected, the id
-            // must match an id of one of the <server>s defined in the settings
-            for (String ap : settings.getActiveProfiles()) {
-                List<Repository> repositories = settings.getProfilesAsMap().get(ap).getRepositories();
-                repositories.forEach(r -> {
-                    if ("test-server".equals(r.getId())) {
-                        r.setUrl("http://localhost:" + localServer.getLocalPort() + "/maven/repository");
-                    }
-                });
-            }
-            for (Mirror mirror : settings.getMirrors()) {
-                if ("test-mirror".equals(mirror.getId())) {
-                    mirror.setUrl("http://localhost:" + localServer.getLocalPort() + "/maven/mirror");
-                }
-            }
-
-            // for test, we'll use hardcoded version
-            String localRepository = "target/.m2/repository";
-            // local repository manager is required and never set implicitly
-            File m2Repo = new File(localRepository);
-            FileUtil.removeDir(m2Repo);
-
-            // we can finally create a session to resolve artifacts
-            RepositorySystemSession session
-                    = downloader.configureRepositorySystemSession(registry, systemProperties, settings, m2Repo);
-
-            List<RemoteRepository> remoteRepositories = downloader.configureRemoteRepositories(settings, null, false);
-
-            // finally we can resolve the artifact
-            ArtifactRequest request = new ArtifactRequest();
-            // repositories from the settings. auth, mirrors and proxies should be handled automatically if we use
-            // one handy method: org.eclipse.aether.RepositorySystem#newResolutionRepositories()
-            for (RemoteRepository r : repositorySystem.newResolutionRepositories(session, remoteRepositories)) {
-                request.addRepository(r);
-            }
-            request.setArtifact(new DefaultArtifact("org.apache.camel", "camel-anything", "jar", "3.42.0"));
-
-            ArtifactResult result = repositorySystem.resolveArtifact(session, request);
-            assertTrue(result.getArtifact().getFile().isFile());
-            assertEquals("/mirror/org/apache/camel/camel-anything/3.42.0/camel-anything-3.42.0.jar",
-                    Files.readString(result.getArtifact().getFile().toPath()));
         }
     }
