+++ parent/pom.xml
@@ -253,6 +253,7 @@
         <jakarta-enterprise-cdi-api-version>4.0.1</jakarta-enterprise-cdi-api-version>
         <jakarta-api-version>2.1.5</jakarta-api-version>
         <jakarta.el-version>3.0.3</jakarta.el-version>
+        <jakarta-inject-version>2.0.1</jakarta-inject-version>
         <jakarta-xml-bind-api-version>4.0.0</jakarta-xml-bind-api-version>
         <jakarta-xml-soap-api-version>3.0.0</jakarta-xml-soap-api-version>
         <jakarta-xml-ws-api-version>4.0.0</jakarta-xml-ws-api-version>
@@ -522,6 +523,11 @@
             <dependency>
                 <groupId>org.apache.camel</groupId>
                 <artifactId>camel-tooling-model</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.camel</groupId>
+                <artifactId>camel-tooling-maven</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
