+++ tooling/camel-tooling-maven/src/main/java/org/apache/camel/tooling/maven/MavenResolutionException.java
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.camel.tooling.maven;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class MavenResolutionException extends Exception {
+
+    private final List<String> repositoriesUsed = new ArrayList<>();
+
+    public MavenResolutionException() {
+    }
+
+    public MavenResolutionException(String message) {
+        super(message);
+    }
+
+    public MavenResolutionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public List<String> getRepositories() {
+        return repositoriesUsed;
+    }
+
+}
