+++ tooling/camel-tooling-maven/src/main/java/org/apache/camel/tooling/maven/package-info.java
@@ -1,0 +1,27 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * <p>
+ * Package with interfaces/classes to be used when proper/canonical Maven resolution is required. Currently it's:
+ * <ul>
+ * <li>catalog/camel-catalog-maven (previously using Ivy/Grape)</li>
+ * <li>dsl/camel-kamelet-main (previously using Shrinkwrap)</li>
+ * </ul>
+ * </p>
+ */
+package org.apache.camel.tooling.maven;
