+++ tooling/camel-tooling-maven/src/test/java/org/apache/camel/tooling/maven/support/DIRegistryTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.main.injection;
+package org.apache.camel.tooling.maven.support;
 
 import java.io.FilenameFilter;
 import java.io.IOException;
