+++ tooling/camel-tooling-maven/src/test/resources/.m2/settings.xml
