+++ tooling/pom.xml
@@ -41,6 +41,7 @@
         <module>camel-tooling-util</module>
         <module>openapi-rest-dsl-generator</module>
         <module>maven</module>
+        <module>camel-tooling-maven</module>
     </modules>
 
     <properties>
